Fix TIM1 timer prescaler is computed incorrectly
Previous implementation had an error in computing the TIM1 prescaler,
though it was not noticable because the prescaler was large. With a
smaller prescaler (7200 -> 72), the off by one causes the TIM1
period to be a few % less than 1 second. This can be seen in the fade
rate of the LED.