I2C and Wiring Changes
In Wiring I moved repeated code to a function to make reading the file
simpler as recommend by @fearless40.

Un-commented out the interrupt handle that was partially written.
Looking into using the STM32 I2C interrupt handler to allow the spark
core to allow multiple master on the bus and give the function of
putting the core in slave mode based on priority level. This might be
more difficult than I though though, I seem to keep being an error so I
din’t push that code.