renamed test-app-1 to testapp1 - easier to type and read
bugfix: for core-v1/testapp, resulting target file was tests/core-v1/testapp/core-v1/testapp.elf. Fixed this by removing the directory prefix in the target name, since this is present in the target dir. This also allows the user to place the target files in a specific directory (otherwise they'd end up in a subdirectory of what the user specifies.)