Limit seed to 4 bytes even on future hardware
There are only 5 bytes left in the credentials,
so an attempt on any future hardware that might support
64-bit integers to use sizeof(seed) would read past
the end of the credentials buffer.

For future reference, there would be very little cryptographic risk
in starting the seed at offset 32 and allowing 8 bytes, duplicating
use of the initial message ID and token bytes as part of the seed.
However, there is no need to do so at this time.