Support for targeting different products via `SPARK_PRODUCT_ID`.
Renamed WLAN to NET in platform so that we can target any networking subsystem. (BLE, Ethernet and other networking technologies shall be put under the NET folder.)

The product ID is mapped to the appropriate architecture, MCU and NET subsystems. Each product produces it's own set of artefacts so there is no issue with potentially getting unclean builds when changing the products ID (as would be the case if the same artefacts were produced.).