debug build is now successful. fixes issue #473. The fix required exporting `snprintf` from system-part2 since this is used in logging messages in part1, and this was directly linking to _malloc_r. This is visible by a linker error `_sbrk` not found. Since we want all memory allocation to take place in part2, it was necessary to export the snprintf function.
Adding a dependency on part2 from part 1 would normally fail, since there is no registered dependency from part1 to part2. (So as part1 is flashed, it is looking for functions in part2 that don't exist yet.) However,  as the new functions are only used in a debug build they will not be part of the released system modules.  Debug build requires a local build, and consequently both system modules are updated at once.