Fix broken build with PRODUCT_ID & PRODUCT_VERSION
Prior to this commit, building for the Photon, with LTO disabled, and
using the `PRODUCT_ID()` and `PRODUCT_VERSION()` macros in one's
application resulted in the error "module info static details not
linked".

This was the result of a failed assertion on line 167 (the last line) of
modules/photon/user-part/linker.ld. The assertion was checking that the
product ID and version were being added to the `.module_info_suffix`
section, defined in build/arm/linker/module_end.ld, as exactly 4 bytes.

For a product ID of 42 at version 3, we expect to see 40 bytes at the
end of the .bin file, with `2a00 0300` immediately prior, 44-41 from the
end. Instead, the assertion was failing because the linker was creating
`ffff 2a00 ffff 0300` 48-41 from the end.

It's a good thing this assertion existed because without it, the product
ID would silently disappear from the binary from the perspective of any
program trying to read it from the binary file. Most notably, the
product creator dashboard would show the user error messages when
uploading binaries because the file would not match the correct product
ID known to the system.

In dynalib/inc/module_info.inc I have deleted what were intended to be
weak symbols for the product ID and version. These were clearly intended
to be overridden by a developer using the `PRODUCT_ID` and
`PRODUCT_VERSION` macros. Assuming this was truly happening, and noting
that it *does* in fact happen with an LTO build, @satishgn and I were
convinced the linker was aligning these symbols as 32-bits wide despite
their declarations as `uint16_t`.

It turns out the real problem was that the weak attribute was having no
effect when LTO was turned off. Both the supposedly weak value `ffff`
and the value passed to the macros were being added to the given linker
section.

There may be a better solution that preserves our expected total section
length, but simply padding the beginning of the .module_info_suffix with
`ffff ffff` ensures that:

- we have correct default values for product ID and version when the
  macros are not used, and
- we find the product ID and version in the expected locations when the
  macros are used.

Fix FIRM-161