Add C++ handlers and handler_data to subscribe
Spark.function, Spark.subscribe and attachInterrupt can now take a pointer to a member and an instance to easily attach a callback to a C++ object.

`spark_subscribe` did not save the data pointer to be passed to the `EventHandler` callback so I implemented the following solution:
* Don't change the signature of any dynalib method
* Save the handler data pointer in `FilteringEventHandler`
* If the data pointer is non-null, call the handler with the extra data pointer (3 arguments)
* If the data pointer is null, call the handler with 2 arguments

This should be compatible with previous code.