Work around a WICED bug crashing when socket connection fails
The reason for the bug is WICED deallocating the connection handler twice.

It occurs because in `WICED/network/LwIP/WICED/tcpip.c:920` the `netconn_delete( socket->conn_handler );` is not followed by `socket->conn_handler = NULL;` as it is in other places in the code for example in `WICED/network/LwIP/WICED/tcpip.c:931`.