DCD implementation for configuration persistence.
The new DCD format is used in both the bootloader and in the electron HAL.

On startup, the firmware automatically upgrades the configuration from the old DCT format to the new DCD format, and also updates the bootloader to use the new format.  This should mean an automatic and invisible update for beta electron users.