Change thread priorities to sensible defaults
The application thread runs at prio 2. The system thread runs at prio 2 instead of prio 9.

New wiring threads run at prio 2 by default but can be set at creation.

The timer thread in WICED is configured to run at prio 2 as well. This should probably be raised to 3 or 4 to be able to preempt the application and system threads.