Improve EEPROM stability and persistence ### Changes * Don't erase a page of flash ON EVERY BOOT to increase Flash life and avoid erasing Flash in every low power situations * Don't format the EEPROM in case both pages are valid, just erase one * Lock the Flash after initialization or variable write
### Background

Bad chargers and depleted batteries cause the EEPROM contents to be lost on the Photon and Electron.
http://community.particle.io/t/eeprom-persistence-issue/16514
https://community.particle.io/t/photon-ran-low-on-battery-now-stuck-in-sos/16710
#704

The root cause is still not clear but one factor making it worse is the fact that the STM EEPROM emulation code (AN3390) erases one page of Flash on every boot. This is unnecessary and aweful for Flash life. Also what happens exactly if the microcontroller looses power during this erase operation? Experience shows that the EEPROM contents will be lost.

This is not a complete fix since an EEPROM erase could still happen during a low battery situation, but it will decrease the occurence of these issues.

### Steps to reproduce EEPROM persistence issue

1. Flash this application code

```
#include "application.h"

// Write something to the EEPROM on demand
int eepromWrite(String value) {
  EEPROM.write(0, 0xA5);
  return 0;
}

void setup()
{
  RGB.control(true);
  if(EEPROM.read(0) == 0xFF) {
    // Red means the EEPROM was wiped
    RGB.color(255, 0, 0);
  } else {
    // Yellow means the EEPROM was preserved
    RGB.color(255, 255, 0);
  }
  Particle.function("eepromWrite", eepromWrite);
}
```

2. Call `particle call device eepromWrite`
3. Reboot the Photon and notice the yellow LED meaning EEPROM contents are saved
4. Unplug the USB
5. Replug it partially, fiddle with it, making sure the white LED turn on and off multiple times and even turns partially on
6. Plug USB firmly
7. Notice the red LED meaning the EEPROM contents were lost. Try again if it doesn't fail on the first try

[Video of failure before the fix](https://youtu.be/YSzFfkvTlE8)
[Video of correct behavior after the fix](https://youtu.be/iKj_9vaid5A)