Implement atomic writes for EEPROM emulation
This is a rework the EEPROM emulation that stores individual bytes in 2 pages of Flash memory.

- Use an invalid flag in MSB of the index (EEPROM virtual address) while writing records to make multi-byte writes all or nothing (atomic)
- Add off device tests for the byte-oriented EEPROM emulation algorithm
- Make EEPROM robust in the presence of resets during flash erase or programming
- Check existing values before writing new values in EEPROM (write is now an update)
- Use same EEPROM code for Core, Photon and Electron
- Increase flash life
- Improve the runtime efficiency of multi-byte reads, writes and page swaps, from O(n^2) to O(n)

User API

- Add `HAL_EEPROM_Get` and `HAL_EEPROM_Put` to do efficient atomic multi-byte reads and writes
- Add `HAL_EEPROM_Clear` to format the EEPROM.
- Add `HAL_EEPROM_Has_PEnding_Erase` and `HAL_EEPROM_Perform_Pending_Erase` to allow users to better schedule erasing the alternate page of EEPROM, since this freezes the microcontroller completely for up to 500ms. If this isn't called in the user code, the page will be erased before the page swap.

Capacity

Core uses 2 pages of 1024 bytes and can store up to 255 bytes of EEPROM.

Photon and Electron use 1 page of 16 Kb and 1 page of 64 Kb and can store up to 4095 bytes of EEPROM.

Performance

Single byte or multi byte reads and writes with a freshly erased EEPROM page take less than 0.1 ms. Reads and writes with a nearly full EEPROM page take 3 ms, regardless of the number of bytes read/written. Previously these operations took 3 ms per byte read/written.

On the Photon/Electron the page swap time went from 1400 ms to 500 ms. If the alternate page is already erased by using `EEPROM.performPendingErase()`, the swap time is shorter at around 15 ms and doesn't freeze the processor for long periods of time.

Because the full 64 Kb page is used on the Photon/Electron, flash life is 2.5x the previous algorithm (20475 byte writes for 2 erases instead of 8190 byte writes for 2 erases).