Blinking Cyan bugfix
The Message is instanciated with a prefix and a suffix to be able to
store the message length in the prefix, and a padding in the suffix.
The prefix is 2 bytes long, and the padding is 16 bytes long.

The buffer was pointing to queue+suffix instead of queue+prefix, leading
to possible corruption when the queue was full (sending more than one
message per second).