Give a compiler error for long names in Particle.variable/function
A common error is to call `Particle.function("mylongfunction", ...)` with a name longer than 12 characters and overlook to check the return value. It will compile and silently ignore to register the function at run time.

This PR adds a compile-time check for the length of the name parameter of `Particle.function()` and `Particle.variable()` when the name is a static string literal (type `char [N]`). If the parameter is a dynamic string (type `char *`) or `String` then no check will be done at compile time.