Limit heap memory usage during page swap to 256 bytes
Batch the reads during page swap as a tradeoff between speed and memory. Caching the address of all records in one read uses up to 4 Kb of heap. Reading and copying one record at a time takes up to 6 seconds. This compromise uses 256 ms and 100 ms on the Photon/Electron.