fixes #896 for G350
U260 and G350 both were recv a +UUSOCL URC for a remotely closed socket. The U260 would receive this later in time and therefor did not see any issue.  The G350 would recv the URC earlier which caused the socketReadable() routine to throw an unexpected error by setting OK to false.  This was keeping the `cnt` from being returned for the last bytes received.  When not `readable` now, we ensure `len` is set to 0 and OK is set to `true`.