Only remake $(TARGET_BASE).elf el al. if necessary
Commit ee1d29e made it possible to run program-dfu and friends without
first running `make all`, but it also made it so the $(TARGET_BASE).elf
recipe is constantly executed, whether it needs it or not.

The problem is that currently, $(MAKE_DEPENDENCIES) is phony, so real
targets shouldn't depend on that [1]. Phony prerequisites are never up
to date, so neither is anything that depends on them, and make rebuilds
anything that isn't up to date.

One solution is to move the phony prerequisites from the real targets to
program-dfu, which is itself phony. That's how the `all` target worked
before ee1d29e.

[1] https://www.gnu.org/software/make/manual/html_node/Phony-Targets