Don't check and write protect the system firmware on boot
There is a possibility that in low power / unreliable power situations that checking if the system firmware flash pages are write protected will give the wrong result. The system firmware would attempt to write protect the system firmware flash pages, which erases the option bytes and if the device resets at that moment, sets the write protection level 1.

A similar change was made to the bootloader in https://github.com/particle-iot/device-os/pull/1578